workers/upscale/requirements.txt
runpod
huggingface_hub
torch>=2.2.0
diffusers>=0.31.0
transformers>=4.51.0
accelerate>=0.30.0
safetensors>=0.4.0
Pillow>=10.0.0
peft>=0.12.0
controlnet-aux
